requirements.txt
streamlit
langchain-community
langchain-core
langchain-groq
chromadb
pandas
python-dotenv
uuid
pysqlite3-binary
hnswlib
sentence-transformers
